feat(apprentissage): retours review
