+++ src/client/components/layouts/Header/Header.test.tsx
@@ -197,7 +197,7 @@ describe('Header', () => {
 			it('ON, affiche le lien vers l’enquête de satisfaction', () => {
 				// GIVEN
 				mockUseRouter({ pathname: '/' });
-				// process.env.NEXT_PUBLIC_CAMPAGNE_APPRENTISSAGE_FEATURE = '1';
+				process.env.NEXT_PUBLIC_CAMPAGNE_APPRENTISSAGE_FEATURE = '1';
 
 				// WHEN
 				render(<Header/>);
@@ -226,6 +226,7 @@ describe('Header', () => {
 				// GIVEN
 				mockUseRouter({ pathname: '/' });
 				process.env.NEXT_PUBLIC_ENQUETE_SATISFACTION_FEATURE = '0';
+				process.env.NEXT_PUBLIC_ENQUETE_SATISFACTION_URL = 'https://docs.google.com/forms/d/e/1FAIpQLSeY3bU5cQlKNCO6B5VRJhPe7j6LwOXLXBikLrzKVAEFkUQPYw';
 
 				// WHEN
 				render(<Header/>);
