+++ src/styles/_utilities.scss
@@ -5,3 +5,8 @@
 @forward "theme/variables";
 @forward "typography/mixins";
 @forward "typography/placeholders";
+
+$_one-rem-in-pixel: 16;
+@function pixel-to-rem($pixel-size) {
+  @return calc($pixel-size / $_one-rem-in-pixel) * 1rem;
+}
